Added weighting on distance to closest bar position based on DISTANCE parameter
  More specifically, we use
likelihoods of note onset events (in
`code.MonoNoteParameters.barPositionDist_Probs`), at different bar
position (e.g. from 0 to 9, depending on the bar). In the original
version the transition likelihood from silence to a following note
attack state is distributed by the pitch difference from current to
following note. NOTE: the sum of the transition likelihoods over all
possible following notes is a constant `1-selfSilenceTransition`. In
this version, on decoding, this constant is replaced for each time frame
by the likelihood of the closest bar position from
`barPositionDist_Probs`. This means essentially, that the same
pitch-difference distribution scheme is kept, but scaled varyingly when
close in time to a beginning of the bar (e.g. scaled more at downbeats
and less else).  
All this is implemented as 2-dimensional array:
`code.MonoNoteHMM.MonoNoteHMM.transProbs`